feat(sidepane): delete old filetree.go and mdtoc.go from ui package
Remove the original filetree.go, filetree_test.go, mdtoc.go, and
mdtoc_test.go files that were ported to app/ui/sidepane/ in earlier
tasks. Relocate isMarkdownFile and isFullContext helper methods to
loaders.go (their only caller). Production code builds cleanly; test
files will be updated in the next task.
